requirements.txt
# --- Core ---
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# --- Market Data (free, no API key needed) ---
yfinance>=0.2.36
ta>=0.11.0               # Technical indicators (replaces pandas-ta; works on Python 3.10+)
                         # pandas-ta is broken on PyPI for Python 3.10

# --- Broker (Alpaca - paper trading is free at alpaca.markets) ---
alpaca-py>=0.21.0
ib-insync>=0.9.86         # Interactive Brokers adapter (optional)

# --- Database ---
sqlalchemy>=2.0.0
pyarrow>=15.0.0

# --- Research data ingestion ---
boto3>=1.34.0

# --- Async ---
aiohttp>=3.9.0
anyio>=4.0.0             # Async test runner (used by pytest via anyio plugin)
websockets>=13.0

# --- CLI & display ---
rich>=13.0.0
click>=8.0.0

# --- Testing ---
pytest>=7.0.0

# --- Research ML (optional) ---
# Only required for the research training pipeline; core runtime does not need them.
xgboost>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
